Conflicts:
	src/add-ons/mail_daemon/inbound_filters/match_header/ConfigView.cpp
	src/add-ons/mail_daemon/inbound_filters/match_header/RuleFilter.cpp
	src/add-ons/mail_daemon/inbound_filters/spam_filter/SpamFilterConfig.cpp
	src/add-ons/mail_daemon/inbound_protocols/imap/imap_lib/IMAPProtocol.cpp
	src/add-ons/mail_daemon/inbound_protocols/pop3/POP3.cpp
	src/kits/mail/MailMessage.cpp
	src/preferences/mail/ConfigWindow.cpp

+++ src/kits/mail/MailMessage.cpp
@@ -129,7 +129,7 @@ BEmailMessage::ReplyMessage(mail_reply_to_mode replyTo, bool accountFromMail,
 		BString cc;
 
 		for (int32 i = list.CountItems(); i-- > 0;) {
-			char* address = (char*)list.RemoveItem(0L);
+			char* address = (char*)list.RemoveItem((int32)0);
 
 			// Add everything which is not the sender and not already in the
 			// list
@@ -737,7 +737,7 @@ BEmailMessage::RenderToRFC822(BPositionIO* file)
 
 	BString recipients;
 	for (int32 i = recipientList.CountItems(); i-- > 0;) {
-		char *address = (char *)recipientList.RemoveItem(0L);
+		char *address = (char *)recipientList.RemoveItem((int32)0);
 
 		recipients << '<' << address << '>';
 		if (i)
